Moved `BlessWindow` class to `duser` app
and rerouted `BlessWindow` import in `main.py`.
